requirements.txt
django==3.2.7
djangorestframework==3.12.4
Pillow==8.3.2
django-filter==21.1
djangorestframework-simplejwt==4.8.0